feat: source test data from .ld-flags.json if exists
This means that if LAUNCHDARKLY_CONFIGURATION is not set, but an
.ld-flags.json file exists in the directory the Go program was invoked
from, the SDK will automatically fall-back to test mode using the JSON
file as the data source.

This provides a transparent way for users to consume the SDK for
local development or testing. They don't need to modify the
initialisation code to explicitly invoke the test mode.