feat: configure big segments and use refined config structure (#38)
The wrapper is now configured to get Big Segments data from the Relay Proxy persistent storage regardless of the connection mode. In order for us to actually support big segments, we need to connect our clients directly to the data table, and that requires DynamoDB GetItem permission on relay table.